+++ src/precheck/package.rs
@@ -1,7 +1,8 @@
-use super::PrecheckError;
+use super::util::passes_spctl;
+use super::Status;
+
 use std::error::Error;
 use std::path::PathBuf;
-use std::process::Command;
 
 pub(super) struct DeveloperIdCheck;
 
@@ -10,21 +11,16 @@ impl super::Precheck for DeveloperIdCheck {
         "Developer ID signing"
     }
 
-    fn run(&self, input_path: &PathBuf) -> Result<(), Box<dyn Error>> {
-        let output = Command::new("/usr/sbin/pkgutil")
-            .arg("--check-signature")
-            .arg(input_path.as_os_str())
-            .output()
-            .unwrap();
-
-        if !output.status.success() {
-            return Err(PrecheckError::new(
+    fn run(&self, input_path: &PathBuf) -> Result<Status, Box<dyn Error>> {
+        // Note: may also use "/usr/sbin/pkgutil --check-signature"
+        if passes_spctl(&vec!["-t", "install"], input_path)? {
+            Ok(Status::Pass)
+        } else {
+            Ok(Status::fail_with(
                 "Package is not signed with a Developer ID certificate.",
                 "Make sure to provide the --sign <installer identity> argument to pkgbuild.",
-            )
-            .into());
+                None,
+            ))
         }
-
-        Ok(())
     }
 }